rag_engine/main.py
import uvicorn
from fastapi import FastAPI
from pydantic import BaseModel
from dotenv import load_dotenv
import os

# Import langchain and openai, even if not used yet
import langchain
import openai

# Load environment variables (e.g., for API keys)
load_dotenv()
# Example: OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

app = FastAPI()

# Pydantic model for the incoming query
class QueryRequest(BaseModel):
    query: str

# Pydantic model for the response (optional, but good practice)
class RAGResponse(BaseModel):
    answer: str
    sources: list

def get_rag_response(query: str) -> dict:
    """
    Simulates the RAG process.
    Returns a hardcoded dictionary with a mock answer and sources.
    """
    return {
        "answer": f"This is a mock RAG response to your query: '{query}'. Actual processing will be implemented later.",
        "sources": [
            {"name": "Mock Source 1.pdf", "page": 1},
            {"name": "Mock Source 2.txt", "snippet": "A relevant snippet from the document..."}
        ]
    }

@app.post("/process_query", response_model=RAGResponse)
async def process_query_endpoint(request: QueryRequest):
    """
    FastAPI POST endpoint that accepts a query, calls get_rag_response,
    and returns the result.
    """
    rag_result = get_rag_response(request.query)
    return rag_result

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
