rag_engine/requirements.txt
langchain
openai
fastapi
uvicorn[standard]
python-dotenv
